Fix: Resolve potential errors
This commit addresses potential errors in the codebase.
