+++ src/components/navigation/Sidebar.tsx
@@ -1,0 +1,137 @@
+
+import React from 'react';
+import { useLocation } from 'react-router-dom';
+import { 
+  Hotel, 
+  CalendarDays, 
+  Clock, 
+  House, 
+  Trash, 
+  LogOut, 
+  ChevronLeft, 
+  ChevronRight,
+  Moon, 
+  Sun,
+} from 'lucide-react';
+import { Button } from '@/components/ui/button';
+import { cn } from '@/lib/utils';
+import { useAuth } from '@/context/AuthContext';
+import NavItem from './NavItem';
+import { useToast } from '@/hooks/use-toast';
+
+interface SidebarProps {
+  collapsed: boolean;
+  toggleSidebar: () => void;
+  theme: 'light' | 'dark';
+  toggleTheme: () => void;
+}
+
+const Sidebar: React.FC<SidebarProps> = ({ collapsed, toggleSidebar, theme, toggleTheme }) => {
+  const location = useLocation();
+  const { signOut } = useAuth();
+  const { toast } = useToast();
+  
+  const navItems = [
+    { 
+      to: '/', 
+      label: 'Dashboard', 
+      icon: <Hotel className="h-4 w-4" /> 
+    },
+    { 
+      to: '/rooms', 
+      label: 'Rooms', 
+      icon: <House className="h-4 w-4" /> 
+    },
+    { 
+      to: '/schedule', 
+      label: 'Schedule', 
+      icon: <CalendarDays className="h-4 w-4" /> 
+    },
+    { 
+      to: '/tasks', 
+      label: 'Tasks', 
+      icon: <Trash className="h-4 w-4" /> 
+    },
+    { 
+      to: '/staff', 
+      label: 'Staff', 
+      icon: <Clock className="h-4 w-4" /> 
+    },
+  ];
+  
+  return (
+    <div 
+      className={cn(
+        "bg-white dark:bg-sidebar shadow-sm border-r dark:border-sidebar-border transition-all duration-300 ease-in-out",
+        collapsed ? "w-16" : "w-64"
+      )}
+    >
+      <div className={cn("p-4 border-b dark:border-sidebar-border flex items-center justify-between", 
+        collapsed ? "flex-col" : "")}>
+        <div className={cn("flex items-center", collapsed ? "flex-col" : "space-x-2")}>
+          <Hotel className="h-6 w-6 text-hotel-primary" />
+          {!collapsed && <h1 className="text-xl font-bold text-hotel-dark dark:text-white">HouseHarmony</h1>}
+        </div>
+        <Button 
+          variant="ghost" 
+          size="sm" 
+          className="ml-auto" 
+          onClick={toggleSidebar}
+        >
+          {collapsed ? <ChevronRight className="h-4 w-4" /> : <ChevronLeft className="h-4 w-4" />}
+        </Button>
+      </div>
+      
+      {!collapsed && (
+        <p className="text-xs text-muted-foreground dark:text-sidebar-foreground/70 px-4 pt-1">Hotel Management System</p>
+      )}
+      
+      <div className="p-4">
+        <nav className="space-y-1">
+          {navItems.map((item) => (
+            <NavItem 
+              key={item.to}
+              to={item.to}
+              label={item.label}
+              icon={item.icon}
+              active={location.pathname === item.to}
+              collapsed={collapsed}
+            />
+          ))}
+        </nav>
+      </div>
+      <div className="p-4 mt-auto border-t dark:border-sidebar-border">
+        <div className="flex mb-2 justify-between">
+          <Button 
+            variant="ghost" 
+            size="sm" 
+            onClick={toggleTheme} 
+            className="dark:text-white"
+          >
+            {theme === 'light' ? <Moon className="h-4 w-4" /> : <Sun className="h-4 w-4" />}
+          </Button>
+          
+          {!collapsed && (
+            <span className="text-xs text-muted-foreground dark:text-sidebar-foreground/70 flex items-center">
+              {theme === 'light' ? 'Light' : 'Dark'} mode
+            </span>
+          )}
+        </div>
+        
+        <Button 
+          variant="outline" 
+          className={cn(
+            "justify-start text-red-500 hover:text-red-600 hover:bg-red-50 dark:hover:bg-red-900/10 dark:border-sidebar-border",
+            collapsed ? "w-full px-2" : "w-full"
+          )}
+          onClick={signOut}
+        >
+          <LogOut className={collapsed ? "" : "mr-2"} size={16} />
+          {!collapsed && "Sign Out"}
+        </Button>
+      </div>
+    </div>
+  );
+};
+
+export default Sidebar;
